requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.26.143
botocore==1.29.143
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.0
cssselect2==0.7.0
Django==4.2.1
django-cors-headers==4.0.0
django-referrer-policy==1.0
django-ses==3.5.0
django-storages==1.13.2
frozenlist==1.3.3
future==0.18.3
google-auth==2.19.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
jmespath==1.0.1
lxml==4.9.2
multidict==6.0.4
oscrypto==1.3.0
Pillow
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.9.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyVimeo==1.1.0
PyYAML==6.0
qrcode==7.4.2
razorpay==1.3.0
reportlab==3.6.13
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tinydb==4.7.1
tuspy==1.0.0
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.16
webencodings==0.5.1
whitenoise==6.4.0
xhtml2pdf==0.2.11
yarl==1.9.2
